Moved spire.util.Opt into the spire-kernel package.
