+++ kernel/shared/src/main/scala_2.10/spire/util/OptVersions.scala
