+++ kernel/shared/src/main/scala_2.11/spire/util/OptVersions.scala
