+++ kernel/shared/src/main/scala_2.12/spire/util/OptVersions.scala
